+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ## 2.7 branch
 ### 2.7.0
 * `Youtube` class does not contain more static methods;
+* `Youtube::getInfo()` method returns an object;
 * renamed repository and package. Now is `me-cms-youtube`.
 
 ## 2.6 branch